Fix update REST API to chat send
채팅 전송 관련 로직 마무리
서버에서 전달 후 처리 작업 이행 예정
