+++ app/src/main/java/com/fitmate/fitmate/domain/model/FitGroup.kt
