+++ app/src/main/java/com/fitmate/fitmate/presentation/ui/mygroup/MyGroupFragment.kt
@@ -2,6 +2,7 @@ package com.fitmate.fitmate.presentation.ui.mygroup
 
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
